[Enhancement] Create EmptyState component
