+++ app/javascript/dashboard/components/widgets/conversation/Message.vue
@@ -1,30 +1,40 @@
 <template>
-  <li :class="alignBubble" v-if="data.attachment || data.content">
+  <li v-if="data.attachment || data.content" :class="alignBubble">
     <div :class="wrapClass">
       <p
-        :class="{ bubble: isBubble, 'is-private': isPrivate }"
         v-tooltip.top-start="sentByMessage"
+        :class="{ bubble: isBubble, 'is-private': isPrivate }"
       >
         <bubble-image
+          v-if="data.attachment && data.attachment.file_type === 'image'"
           :url="data.attachment.data_url"
           :readable-time="readableTime"
-          v-if="data.attachment && data.attachment.file_type==='image'"
         />
         <bubble-audio
+          v-if="data.attachment && data.attachment.file_type === 'audio'"
           :url="data.attachment.data_url"
           :readable-time="readableTime"
-          v-if="data.attachment && data.attachment.file_type==='audio'"
         />
         <bubble-map
+          v-if="data.attachment && data.attachment.file_type === 'location'"
           :lat="data.attachment.coordinates_lat"
           :lng="data.attachment.coordinates_long"
           :label="data.attachment.fallback_title"
           :readable-time="readableTime"
-          v-if="data.attachment && data.attachment.file_type==='location'"
         />
-        <i class="icon ion-person" v-if="data.message_type === 2"></i>
-        <bubble-text v-if="data.content" :message="message" :readable-time="readableTime"/>
-        <i class="icon ion-android-lock" v-if="isPrivate" v-tooltip.top-start="toolTipMessage" @mouseenter="isHovered = true" @mouseleave="isHovered = false"></i>
+        <i v-if="data.message_type === 2" class="icon ion-person"></i>
+        <bubble-text
+          v-if="data.content"
+          :message="message"
+          :readable-time="readableTime"
+        />
+        <i
+          v-if="isPrivate"
+          v-tooltip.top-start="toolTipMessage"
+          class="icon ion-android-lock"
+          @mouseenter="isHovered = true"
+          @mouseleave="isHovered = false"
+        ></i>
       </p>
     </div>
     <!-- <img v-if="showSenderData" src="https://chatwoot-staging.s3-us-west-2.amazonaws.com/uploads/avatar/contact/3415/thumb_10418362_10201264050880840_6087258728802054624_n.jpg?X-Amz-Algorithm=AWS4-HMAC-SHA256&amp;X-Amz-Credential=AKIAI3KBM2ES3VRHQHPQ%2F20170422%2Fus-west-2%2Fs3%2Faws4_request&amp;X-Amz-Date=20170422T075421Z&amp;X-Amz-Expires=604800&amp;X-Amz-SignedHeaders=host&amp;X-Amz-Signature=8d5ff60e41415515f59ff682b9a4e4c0574d9d9aabfeff1dc5a51087a9b49e03" class="sender--thumbnail"> -->
@@ -46,8 +56,13 @@ export default {
     BubbleMap,
     BubbleAudio,
   },
-  props: ['data'],
   mixins: [timeMixin],
+  props: {
+    data: {
+      type: Object,
+      required: true,
+    },
+  },
   data() {
     return {
       isHovered: false,
@@ -71,11 +86,16 @@ export default {
       return this.data.private;
     },
     toolTipMessage() {
-      return this.data.private ? { content: this.$t('CONVERSATION.VISIBLE_TO_AGENTS'), classes: 'top' } : false;
+      return this.data.private
+        ? { content: this.$t('CONVERSATION.VISIBLE_TO_AGENTS'), classes: 'top' }
+        : false;
     },
     sentByMessage() {
-      return this.data.message_type === 1 && !this.isHovered && this.data.sender !== undefined ?
-        { content: `Sent by: ${this.data.sender.name}`, classes: 'top' } : false;
+      return this.data.message_type === 1 &&
+        !this.isHovered &&
+        this.data.sender !== undefined
+        ? { content: `Sent by: ${this.data.sender.name}`, classes: 'top' }
+        : false;
     },
     wrapClass() {
       return {
@@ -83,14 +103,16 @@ export default {
         'activity-wrap': !this.isBubble,
       };
     },
-
   },
   methods: {
     getEmojiSVG,
     linkify(text) {
       if (!text) return text;
       const urlRegex = /(https?:\/\/[^\s]+)/g;
-      return text.replace(urlRegex, url => `<a href="${url}" target="_blank">${url}</a>`);
+      return text.replace(
+        urlRegex,
+        url => `<a href="${url}" target="_blank">${url}</a>`
+      );
     },
   },
 };
